refactor: move activity context related code close to where it's used - Moves provider, context and hook close to an only place where they are   used - Removes extra mocked test and adds another test case for business   logic of useCurrentActivity - refactor type reference
