+++ spec/mocks/index.ts
@@ -1,2 +1,1 @@
 export * from './activities'
-export * from './data'
