+++ src/components/Activities/context/index.ts
@@ -1,0 +1,2 @@
+export * from './ActivityContext'
+export { ActivityProvider } from './ActivityProvider'
