+++ src/components/Activities/hooks/index.ts
@@ -1,2 +1,1 @@
-export { ActivityProvider } from './ActivityProvider'
 export { useCurrentActivity } from './useCurrentActivity'
