+++ src/components/Activities/hooks/useCurrentActivity.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import { ActivityContext, ActivityContextInterface } from './ActivityContext'
+import { ActivityContext, ActivityContextInterface } from '../context'
 
 export const useCurrentActivity = (): ActivityContextInterface =>
   useContext(ActivityContext)
