+++ src/components/Activities/types.ts
@@ -1,2 +1,5 @@
-export { ActivityObjectType } from '../../types/generated/types-orchestration'
+export {
+  ActivityObjectType,
+  ActivityStatus,
+} from '../../types/generated/types-orchestration'
 export type { Activity } from '../../types'
