--- src/hooks/activityNavigation/types.ts
@@ -1,1 +1,0 @@
-export type { Activity } from '../../types'
