+++ src/hooks/useForm/index.ts
@@ -1,3 +1,2 @@
 export { useForm } from './useForm'
 export type { Form } from './types'
-export { ActivityStatus } from './types'
